+++ chekr/distchk.c
@@ -14,7 +14,7 @@ int *imod;
 int *ierr;
 int icherr[5];
 {
-  long ip[5];                           /* ipc parameters */
+  int ip[5];                           /* ipc parameters */
   int ind;
   struct req_rec request;          /* mcbcn request record */
   struct req_buf buffer;           /* mcbcn request buffer */
