+++ class/class2.c
@@ -16,7 +16,7 @@ extern struct fscom *shm_addr;
 #define MAX_TEXT  120
 
 struct  cls_buf {
-	long	mtype;
+	int	mtype;
         struct {
           int     parm[2];
           char    mtext[ MAX_TEXT];
@@ -28,7 +28,7 @@ static int msqid;
 main()
 {
 key_t key;
-long    class;
+int    class;
 struct  cls_buf msg;
 void sem_take(), sem_put();
 int nchars, icount;
